+++ interfaces/freeswitch/number_lookup.py
@@ -1,0 +1,21 @@
+#!/usr/bin/python
+
+import os
+import subprocess
+from freeswitch import *
+
+grab_number = os.path.join(os.environ["PYTHONPATH"], "grab_number", "grab_number.py")
+
+def handler(session, args):
+
+        caller = session.getVariable("caller_id_name")
+        st_lookup=subprocess.check_output([grab_number, "--LogLevel", "X", "query", "--number", caller])
+        st_lookup=st_lookup.strip()
+        session.execute("set","effective_caller_id_name="+st_lookup)
+
+def fsapi(session, stream, env, args):
+
+        st_lookup=subprocess.check_output([grab_number, "query", "--number", args])
+	stream.write(args + " -> " + st_lookup)
+
+